backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pymongo==4.4.0
flask-cors==4.0.1
